@@ -12,34 +12,34 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.1.0",
-    "@angular/common": "^6.1.0",
-    "@angular/compiler": "^6.1.0",
-    "@angular/core": "^6.1.0",
-    "@angular/forms": "^6.1.0",
-    "@angular/http": "^6.1.0",
-    "@angular/platform-browser": "^6.1.0",
-    "@angular/platform-browser-dynamic": "^6.1.0",
-    "@angular/router": "^6.1.0",
+    "@angular/animations": "^7.2.1",
+    "@angular/common": "^7.2.1",
+    "@angular/compiler": "^7.2.1",
+    "@angular/core": "^7.2.1",
+    "@angular/forms": "^7.2.1",
+    "@angular/http": "^7.2.1",
+    "@angular/platform-browser": "^7.2.1",
+    "@angular/platform-browser-dynamic": "^7.2.1",
+    "@angular/router": "^7.2.1",
     "@ngx-translate/core": "^10.0.2",
     "bootstrap": "^3.3.7",
     "core-js": "^2.5.4",
     "moment": "^2.22.2",
-    "ng-http-loader": "^3.1.2",
+    "ng-http-loader": "^5.0.0",
     "ngx-bootstrap": "^3.0.1",
     "ngx-infinite-scroll": "^7.0.1",
     "ngx-moment": "^3.0.1",
     "ngx-pagination": "^3.2.1",
     "ngx-toastr": "^9.1.0",
-    "rxjs": "~6.2.0",
+    "rxjs": "~6.3.3",
     "rxjs-compat": "^6.3.3",
-    "zone.js": "~0.8.26"
+    "zone.js": "~0.8.28"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.8.0",
     "@angular/cli": "^7.2.2",
-    "@angular/compiler-cli": "^6.1.0",
-    "@angular/language-service": "^6.1.0",
+    "@angular/compiler-cli": "^7.2.1",
+    "@angular/language-service": "^7.2.1",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
@@ -55,6 +55,7 @@
     "protractor": "^5.4.1",
     "ts-node": "~3.2.0",
     "tslint": "~5.11.0",
-    "typescript": "~2.9.2"
+    "typescript": "~3.2.4",
+    "webpack": "4.24.0"
   }
 }
